skills/asset-gen/tools/requirements.txt
google-genai
requests
numpy
pillow
scipy
rembg
pymatting
onnxruntime
